fix: auth키 노출문제로 과거 커밋기록으로 이동

+++ src/main/java/org/team/bookshop/domain/order/Service/OrderService.java
@@ -1,7 +1,7 @@
 package org.team.bookshop.domain.order.Service;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.Comparator;
 import java.util.List;
 import java.util.UUID;
@@ -11,8 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.team.bookshop.domain.order.dto.OrderAbstractResponse;
-import org.team.bookshop.domain.order.dto.OrderAddressCreate;
-import org.team.bookshop.domain.order.dto.OrderAddressUpdate;
 import org.team.bookshop.domain.order.dto.OrderCreateRequest;
 import org.team.bookshop.domain.order.dto.OrderItemRequest;
 import org.team.bookshop.domain.order.dto.OrderListResponse;
@@ -26,11 +24,10 @@ import org.team.bookshop.domain.order.enums.OrderStatus;
 import org.team.bookshop.domain.order.repository.DeliveryRepository;
 import org.team.bookshop.domain.order.repository.OrderItemRepository;
 import org.team.bookshop.domain.order.repository.OrderRepository;
-import org.team.bookshop.domain.payment.entity.Payment;
+import org.team.bookshop.domain.payment.entity.Payments;
 import org.team.bookshop.domain.payment.repository.PaymentRepository;
 import org.team.bookshop.domain.product.entity.Product;
 import org.team.bookshop.domain.product.repository.ProductRepository;
-import org.team.bookshop.domain.user.entity.Address;
 import org.team.bookshop.domain.user.entity.User;
 import org.team.bookshop.domain.user.repository.UserRepository;
 import org.team.bookshop.global.error.ErrorCode;
@@ -132,7 +129,7 @@ public class OrderService {
       orderItemRepository.save(orderItem);
       order.getOrderItems().add(orderItem);
     }
-
+    totalPrice = (int) (totalPrice * 0.9);
     order.setOrderTotalAmount(totalCount);
     order.setOrderTotalPrice(totalPrice);
 
@@ -149,52 +146,87 @@ public class OrderService {
     return order.getId();
   }
 
+  // 주문 수정기능
+//  @Transactional
+//  public OrderResponse update(OrderUpdateRequest orderUpdateRequest) {
+//    String merchantUid = orderUpdateRequest.getMerchantUid();
+//
+//    OrderAddressUpdate orderAddressUpdate = orderUpdateRequest.getOrderAddressUpdate();
+//
+//    Order order = orderRepository.findByMerchantUid(merchantUid)
+//        .orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER));
+//    Delivery delivery = order.getDelivery();
+//    Address address = delivery.getAddress();
+//
+//    // 주문 수정 요청을 바탕으로 기존의 주소를 수정
+//    address.setZipcode(orderAddressUpdate.getZipcode());
+//    address.setAddress1(orderAddressUpdate.getAddress1());
+//    address.setAddress2(orderAddressUpdate.getAddress2());
+//    address.setRecipientName(orderAddressUpdate.getRecipientName());
+//    address.setRecipientPhone(orderAddressUpdate.getRecipientPhone());
+//    return OrderResponse.builder()
+//        .orderId(order.getId())
+//        .orderStatus(OrderStatus.PAYED)
+//        .merchantUid(order.getMerchantUid())
+//        .orderTotalAmount(order.getOrderTotalPrice())
+//        .orderTotalPrice(order.getOrderTotalPrice())
+//        .build();
+//  }
+
+  // 주문의 주소정보를 업데이트
   @Transactional
-  public OrderResponse update(OrderUpdateRequest orderUpdateRequest) {
+  public String update(OrderUpdateRequest orderUpdateRequest) {
+
     String merchantUid = orderUpdateRequest.getMerchantUid();
 
-    OrderAddressUpdate orderAddressUpdate = orderUpdateRequest.getOrderAddressUpdate();
-
+    // 전달된 merchantUid에 해당하는 order를 조회
     Order order = orderRepository.findByMerchantUid(merchantUid)
         .orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER));
-    Delivery delivery = order.getDelivery();
-    Address address = delivery.getAddress();
 
-    // 주문 수정 요청을 바탕으로 기존의 주소를 수정
-    address.setZipcode(orderAddressUpdate.getZipcode());
-    address.setAddress1(orderAddressUpdate.getAddress1());
-    address.setAddress2(orderAddressUpdate.getAddress2());
-    address.setRecipientName(orderAddressUpdate.getRecipientName());
-    address.setRecipientPhone(orderAddressUpdate.getRecipientPhone());
-    return OrderResponse.builder()
-        .orderId(order.getId())
-        .orderStatus(OrderStatus.PAYED)
-        .merchantUid(order.getMerchantUid())
-        .orderTotalAmount(order.getOrderTotalPrice())
-        .orderTotalPrice(order.getOrderTotalPrice())
-        .build();
+    // 결제를 하여야만 배송정보를 업데이트 가능
+    if (order.getOrderStatus() == OrderStatus.ACCEPTED) {
+      throw new ApiException(ErrorCode.CANNOT_UPDATE_ORDER);
+    }
+
+    // 해당 order의 배송정보 조회
+    Delivery delivery = order.getDelivery();
+
+    // 해당 배송정보를 업데이트
+    delivery.updateAddressByOrderUpdateRequest(orderUpdateRequest.getOrderAddressUpdate());
+
+    return merchantUid;
   }
+
 
   @Transactional
   public Long delete(Long orderId) {
     Order order = orderRepository.findWithAllRelatedEntityById(orderId);
+    Delivery delivery = order.getDelivery();
 
-    // 해당 주문 내에 포함된 상품의 재고를 다시 원상복구 한다.
+    // 만약 order의 delivery상태가 "출발"이거나 그 이상의 단계라면 주문 취소는 불가능하다.
+    if(delivery.getDeliveryStatus() !=  DeliveryStatus.READY)
+      throw new ApiException(ErrorCode.CANNOT_CANCEL_ORDER);
+
+    // 주문 품목 취소과정
     List<OrderItem> orderItems = order.getOrderItems();
     for (OrderItem orderItem : orderItems) {
-      // 재고수량 회복
+      // 주문한 물품들의 수량을 하나씩 올려준다.
       orderItem.getProduct().increaseStock(orderItem.getOrderCount());
 
-      // orderItem 삭제
+      // 해당 orderItem데이터를 삭제한다
       orderItemRepository.delete(orderItem);
     }
 
-    // 관련된 엔티티인 delivery, address, orderItem모두 삭제
-    Delivery delivery = order.getDelivery();
-    deliveryRepository.delete(delivery);
-    orderRepository.deleteById(orderId);
+    // 관련 payment 삭제
+    Payments payments = paymentRepository.findPaymentsByOrder(order)
+        .orElseThrow(() -> new ApiException(ErrorCode.NO_PAYMENT_INFO_WITH_ORDER));
+    paymentRepository.delete(payments);
 
-    // 삭제된 orderId반환
+    // order정보 삭제
+    orderRepository.delete(order);
+
+    // delivery정보 삭제
+    deliveryRepository.delete(delivery);
     return orderId;
   }
 
@@ -202,20 +234,20 @@ public class OrderService {
     return orderRepository.findWithOrderItems(orderId);
   }
 
-  @Transactional
-  public void setOrderAddress(Long orderId, OrderAddressCreate orderAddressCreate) {
-    Order order = orderRepository.findById(orderId)
-        .orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER));
-
-    Address address = orderAddressCreate.toEntity();
-    Delivery delivery = Delivery.createDelivery(DeliveryStatus.READY, LocalDate.now(), 1L);
-    delivery.setAddress(address);
-
-    deliveryRepository.save(delivery);
-
-    order.setDelivery(delivery);
-
-  }
+//  @Transactional
+//  public void setOrderAddress(Long orderId, OrderAddressCreate orderAddressCreate) {
+//    Order order = orderRepository.findById(orderId)
+//        .orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER));
+//
+//    Address address = orderAddressCreate.toEntity();
+//    Delivery delivery = Delivery.createDelivery(DeliveryStatus.READY, LocalDate.now(), 1L);
+//    delivery.setAddress(address);
+//
+//    deliveryRepository.save(delivery);
+//
+//    order.setDelivery(delivery);
+//
+//  }
 
   // orderListResponse용 : userID 이용 하여 orderListResponse용 dto를 반환하는 기능
   public OrderListResponse findByUserIdForOrderListResponse(Long userId) {
@@ -254,7 +286,7 @@ public class OrderService {
     return order.getOrderTotalPrice() == totalPrice;
   }
 
-  //// 주문을 상세 조회하기 위한 DTO를 반환하는 기능
+  // 주문을 상세 조회하기 위한 DTO를 반환하는 기능
   public OrderResponse getOrderDetail(String merchantUid) {
 
     // 1. 주문 조회하기
@@ -264,16 +296,21 @@ public class OrderService {
 
     // 2. 결제 조회하기 by Order
 
-    Payment payment = paymentRepository.findPaymentByOrder(order)
+    Payments payment = paymentRepository.findPaymentsByOrder(order)
         .orElseThrow(() -> new ApiException(ErrorCode.NO_PAYMENT_INFO_WITH_ORDER));
 
     // 3. 조회한 주문과 결제정보를 바탕으로, 주문 상세조회용 OrderResponse만들기
+    // 반환될 주문 생성일 정보를 문자열로 반환하기 위한 포매터 정의
+    DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy.MM.dd HH : mm : ss");
+
 
     OrderResponse orderResponse = new OrderResponse(
         order.getId(),
         order.getMerchantUid(),
-        order.getOrderItems().stream().map(OrderItem::toOrderItemResponse).collect(Collectors.toList()),
+        order.getOrderItems().stream().map(OrderItem::toOrderItemResponse)
+            .collect(Collectors.toList()),
         order.getOrderStatus(),
+        order.getOrderDateTime().format(formatter),
         order.getDelivery().toOrderDeliveryResponse(),
         order.getOrderTotalPrice(),
         order.getOrderTotalAmount(),
@@ -286,10 +323,12 @@ public class OrderService {
   }
 
   public Order findByMerchantUid(String merchantUid) {
-    return orderRepository.findByMerchantUid(merchantUid).orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER));
+    return orderRepository.findByMerchantUid(merchantUid)
+        .orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER));
   }
 
   public int getOrderTotalPriceByMerchantUid(String merchantUid) {
-    return orderRepository.findByMerchantUid(merchantUid).orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER)).getOrderTotalPrice();
+    return orderRepository.findByMerchantUid(merchantUid)
+        .orElseThrow(() -> new ApiException(ErrorCode.NO_EXISTING_ORDER)).getOrderTotalPrice();
   }
 }
